Crate rust_polynomes

This crate provides a framework for working with polynomes

The most important structure is polynomes::TypedPolynome, the rest of the crate are helpers and wrappers to support it

Usage

use rust_polynomes::{Coeff};
use rust_polynomes::variables::{X, Y};
use rust_polynomes::polynomes::TypedPolynome;

let polynome = Coeff(2u32) * (X + Y) + X * Y;
assert_eq!(polynome.substitute(vec![
    (X, 2u32),
    (Y, 4u32)
]).unwrap(), 20u32);

Custom coefficient data types

You can use this crate with custom types of your own - all you need to do is meet the requirements for Semiring trait described in traits module

Modules

monomes

This module provides structs to describe monomes

polynomes

This module provides structs to describe polynomes

traits

This crate provides a Semiring trait that is used as a constraint for coefficient types

variables

This module provides object variables that can be used to describe the polynome

Structs

Coeff

This struct is a wrapper for coefficients. When ,ultiplied with variables and monomes it results in TypedMonome or TypedPolynome

Enums

SubstitutionError

This enum type describes errors that can occur when calling substitue method