Struct rust_polynomes::polynomes::TypedPolynome

pub struct TypedPolynome<T: Semiring> {
    pub monomes: Vec<TypedMonome<T>>,
}

This structure represents a polynome with a fixed coeffficient type. You can do all types of algebraic operations with it, substitute values or even substitute another TypedMonome instead of variables

Usage

use rust_polynomes::{Coeff};
use rust_polynomes::variables::{X, Y};
use rust_polynomes::polynomes::TypedPolynome;

let polynome = Coeff(2u32) * (X + Y) + X * Y;
assert_eq!(polynome.substitute(vec![
    (X, 2u32),
    (Y, 4u32)
]).unwrap(), 20u32);

You can substitute variables with polynomes like this: Typed monome substitution

use rust_polynomes::{Coeff, SubstitutionError};
use rust_polynomes::variables::{X, Y};
use rust_polynomes::monomes::{UntypedMonome, TypedMonome};
use rust_polynomes::polynomes::{UntypedPolynome, TypedPolynome};

let polynome = Coeff(2u32) * X + Coeff(3u32) * Y;
assert_eq!(polynome.substitute(vec![
    (X, 3u32),
    (Y, 2u32)
]).unwrap(), 12u32);

You can substitute another polynome instead of a variable like this:

use rust_polynomes::{Coeff, SubstitutionError};
use rust_polynomes::variables::{X, Y, Z};
use rust_polynomes::monomes::{UntypedMonome, TypedMonome};
use rust_polynomes::polynomes::{UntypedPolynome, TypedPolynome};
use num_traits::pow::Pow;

let polynome = Coeff(2u32) * X * Y + Coeff(3u32) * Y;

let mut polynome : TypedPolynome<i32> = X.pow(3).into();
polynome.substitute_polynome(X, (X + Y).into());
let mut expected = X.pow(3) + Coeff(3i32) * (X * X * Y + X * Y * Y) + Y.pow(3);
expected.order();

assert_eq!(polynome, expected);

Fields
monomes: Vec<TypedMonome<T>>
Implementations

impl<T: Semiring> TypedPolynome<T>

pub fn new(val: T) -> Self

Constant polynome

Examples

use rust_polynomes::polynomes::TypedPolynome;

let polynome = TypedPolynome::<f32>::new(6.5f32);
assert_eq!(polynome.monomes.len(), 1usize);
assert_eq!(polynome.monomes[0].coeff, 6.5f32);
assert_eq!(polynome.monomes[0].vars.powers.len(), 0);

pub fn substitute<U: Mul<Output = U> + One + Sized + Clone>( &self, substitute_list: Vec<(Var, U)>, ) -> Result<<T as Mul<U>>::Output, SubstitutionError>
where T: Mul<U>, <T as Mul<U>>::Output: Add<Output = <T as Mul<U>>::Output> + Zero,

Typed monome substitution

Examples

use rust_polynomes::{Coeff, SubstitutionError};
use rust_polynomes::variables::{X, Y};
use rust_polynomes::monomes::{UntypedMonome, TypedMonome};
use rust_polynomes::polynomes::{UntypedPolynome, TypedPolynome};

let polynome = Coeff(2u32) * X + Coeff(3u32) * Y;
assert_eq!(polynome.substitute(vec![
    (X, 3u32),
    (Y, 2u32)
]).unwrap(), 12u32);

assert_eq!(polynome.substitute(
    vec![(X, 1u32)]
), Err(SubstitutionError::MissingVariable(Y)));

assert_eq!(polynome.substitute(
    vec![
        (X, 1u32),
        (Y, 1u32),
        (Y, 1u32),
    ]
), Err(SubstitutionError::RepeatingVariable(Y)));

pub fn substitute_polynome(&mut self, var: Var, polynome: TypedPolynome<T>)

Typed monome substitution

Examples

use rust_polynomes::{Coeff, SubstitutionError};
use rust_polynomes::variables::{X, Y, Z};
use rust_polynomes::monomes::{UntypedMonome, TypedMonome};
use rust_polynomes::polynomes::{UntypedPolynome, TypedPolynome};
use num_traits::pow::Pow;

let polynome = Coeff(2u32) * X * Y + Coeff(3u32) * Y;

let mut first = polynome.clone();
first.substitute_polynome(Y, Z.into());
assert_eq!(first, Coeff(2u32) * X * Z + Coeff(3u32) * Z);

let reserve = first.clone();
first.substitute_polynome(Y, Z.into());
assert_eq!(first, reserve);

let mut second : TypedPolynome<i32> = X.pow(3).into();
second.substitute_polynome(X, (X + Y).into());
let mut expected = X.pow(3) + Coeff(3i32) * (X * X * Y + X * Y * Y) + Y.pow(3);
expected.order();

assert_eq!(second, expected);

pub fn order(&mut self)

Examples

use rust_polynomes::Coeff;
use rust_polynomes::variables::{X, Y};
use rust_polynomes::monomes::{UntypedMonome, TypedMonome};
use rust_polynomes::polynomes::{UntypedPolynome, TypedPolynome};
use num_traits::pow::Pow;

let mut polynome = X + Y - (Coeff(1i32) * X).into() + Coeff(3i32) * Y;
polynome.order();

assert_eq!(polynome.monomes, vec![
    TypedMonome {
        coeff: 4i32,
        vars: Y.into(),
   }
]);

let mut another = (Coeff(1u32) * X + Y).pow(2);
another.order();
assert_eq!(another.monomes, vec![
    TypedMonome {
        coeff: 2u32,
        vars: X * Y,
    },
    TypedMonome {
        coeff: 1u32,
        vars: X * X,
    },
    TypedMonome {
        coeff: 1u32,
        vars: Y * Y,
    }
]);